feat(constellation): construct a focus chain explicitly before broadcasting focus messages
N.B. This phrase "focus messages" refers to `ConstellationControlMsg::{
FocusDocument, FocusIFrame}`, not `ConstellationControlMsg::
SystemFocus`.

This commit updates `handle_focus_msg` to construct the old/new focus
chains explicitly before sending out focus messages and use those to
decide which pipelines to send the messages to.

One benefit is being able to skip the message submission to pipelines
whose focus states are not affected. Consider the following pipeline
tree:

     (1) top-level document
      |  (1.1) <iframe>
      |        (2) nested document
      |         |  (2.1) <iframe>
      |         |         (3) nested document  <-- focused
      |         |
      |         |  (2.2) <iframe>
      |         |         (4) nested document
      |         |          |
      |         |          |  (4.1) <iframe>
      |         |          |         (5) nested document
      |

(3) is currently focused. Assuming we have system focus, the focus chain
of the top-level BC is `[(3), (2.1), (2), (1.1), (1)]`. Let's suppose
author code executed `frame.focus()` where `frame` refers to (4.1). This
will cause (4) to send `Focus(Some((5)), _)` to the constellation.

(The above description ignores the distinction between pipelines and
browsing contexts.)

Previously, upon receiving this message, the constellation would have
sent `FocusIFrame((1), (2), _)` to (1). Since (2)'s iframe (1.1) is
already focused, this is wasted work.

After this commit, `handle_focus_msg` will explicitly construct the old
and new focus chains, only considering documents. For the above example,
the constructed focus chains look like the following:

 - old_focus_chain_pipelines: [(3), (2), (1)]
 - new_focus_chain_pipelines: [(5), (4), (2), (1)]

`handle_focus_msg` then compares these and sends the messages necessary
to complete the focus operation. In this case, there are only two such
messages: `FocusIFrame((2), (4), _)` to (2) and `FocusDocument((5), _
)` to (5).

Actually, this is not the primary motivation of this change. We want to
support sending "unfocus" messages to unfocused documents ((3) in the
above example) in the future, and constructing a full focus chain is a
crucial step in this. (I.e., this change is intended to be a refactoring
change, except it actually changes an internal behavior.)